% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmytable.R
\name{ggmytable}
\alias{ggmytable}
\title{ggmytable}
\usage{
ggmytable(df, x, y, z)
}
\arguments{
\item{df}{a data frame}

\item{x}{a qualitative vector from df}

\item{y}{a qualitative vector from df}

\item{z}{a quantitative vector from df}
}
\value{
Returns a bar-plot and summary of the selected vectors from df
}
\description{
A short function which produces a ggplot of the data.#'
}
\examples{
\dontrun{ggmytable(mtbe, "MTBE-Detect", "Aquifier", "MTBE-Level")}
}
